requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
celery>=5.3.0
kombu>=5.3.0
django-celery-results>=2.5.0
python-decouple>=3.8
drf-yasg>=1.21.7
pytest>=7.4.0
pytest-django>=4.7.0
flake8>=6.1.0
